+++ aws_secrets/config/config_reader.py
@@ -1,5 +1,6 @@
 import os
 from pathlib import Path
+from typing import Any, Dict
 
 import yaml
 from aws_secrets.config.providers import BaseProvider
@@ -10,6 +11,24 @@ from aws_secrets.miscellaneous import session
 
 
 class ConfigReader(object):
+    """
+        Config Reader handles the configuration and the secrets YAML files
+
+        Also, abstract the encrypt/decrypt function using the Providers
+
+        Args:
+            config_file (`str`): configuration file path
+
+        Attributes:
+            data (`Dict[str, Any]`): configuration parsed YAML
+            secrets_file_path (`str`): secrets YAML file path
+            global_tags (`Dict[str, str]`): map of global tags of the config file
+            secrets_data (`Dict[str, Any]`): secrets parsed YAML
+            session (`Session`): boto3 session object
+            kms_arn (`str`): Main Kms ARN
+            providers (`Dict[str, BaseProvider]`): providers map
+    """
+
     def __init__(self, config_file: str) -> None:
         super().__init__()
         self.config_file = config_file
@@ -26,11 +45,27 @@ class ConfigReader(object):
         self.providers = self.load_providers()
 
     def get_provider(self, name: str) -> BaseProvider:
+        """
+            Get provider by name
+
+            Supported names:
+            - secrets
+            - parameters
+
+            Args:
+                name (`str`): provider name
+
+            Returns:
+                `BaseProvider`: provider object
+        """
         return self.providers[name]
 
-    def _get_secrets_path(self):
+    def _get_secrets_path(self) -> str:
         """
             Get Secret file path based on the environment file.
+
+            Returns:
+                `str`: secrets YAML file path
         """
         if 'secrets_file' not in self.data:
             config_file_path = Path(self.config_file)
@@ -42,7 +77,13 @@ class ConfigReader(object):
 
         return self.data['secrets_file']
 
-    def load_providers(self):
+    def load_providers(self) -> Dict[str, BaseProvider]:
+        """
+            Load providers
+
+            Returns:
+                `Dict[str, BaseProvider]`: providers
+        """
         providers = {
             'secrets': SecretsManagerProvider(self),
             'parameters': SSMProvider(self)
@@ -50,27 +91,52 @@ class ConfigReader(object):
 
         return providers
 
-    def get_kms_arn(self):
+    def get_kms_arn(self) -> str:
+        """
+            Get main KMS Arn
+
+            Returns:
+                `str`: KMS ARN
+        """
         return str(self.data['kms']['arn'])
 
-    def load_secrets_config(self):
+    def load_secrets_config(self) -> Dict[str, Any]:
+        """
+            Load secrets.yaml file to a dict object
+
+            Returns:
+                `Dict[str, Any]`: secrets dict object
+        """
         secrets_data = {}
 
         if os.path.exists(self.secrets_file_path):
             with open(self.secrets_file_path, 'r') as secrets:
                 secrets_data = yaml.safe_load(secrets.read())
+
         return secrets_data
 
-    def load_config(self):
+    def load_config(self) -> Dict[str, Any]:
+        """
+            Load config YAML file to a dict object
+
+            Returns:
+                `Dict[str, Any]`: config dict object
+        """
         with open(self.config_file, 'r') as source:
             data = yaml.safe_load(source.read())
         return data
 
-    def decrypt(self):
+    def decrypt(self) -> None:
+        """
+            Call decrypt function in all the providers
+        """
         self.get_provider('secrets').decrypt()
         self.get_provider('parameters').decrypt()
 
-    def encrypt(self):
+    def encrypt(self) -> None:
+        """
+            Call encrypt function in all the providers and delete the `value` property from the secret entries
+        """
         self.secrets_data = {
             'secrets': self.get_provider('secrets').encrypt(),
             'parameters': self.get_provider('parameters').encrypt()
@@ -78,7 +144,13 @@ class ConfigReader(object):
 
         self._delete_plain_text_property()
 
-    def _delete_plain_text_property(self):
+    def _delete_plain_text_property(self) -> None:
+        """
+            Delete `value` property from the secrets entries
+
+            For AWS Secrets Manager all the entries
+            For AWS SSM Parameter all the parameter with `SecureString` type
+        """
         all_sensetive_items = self.get_provider('secrets').get_sensible_entries() + \
             self.get_provider('parameters').get_sensible_entries()
 
@@ -86,7 +158,10 @@ class ConfigReader(object):
             if 'value' in item and (isinstance(item['value'], str) or isinstance(item['value'], dict)):
                 del item['value']
 
-    def save(self):
+    def save(self) -> None:
+        """
+            Save config and secrets changes in the disk
+        """
         with open(self.config_file, 'w') as outfile:
             yaml.safe_dump(self.data, outfile)