+++ aws_secrets/config/providers/ssm/provider.py
@@ -1,14 +1,45 @@
+from typing import Any, Dict, List, Optional
+
 import click
-from aws_secrets.config.providers import BaseEntry, BaseProvider
+from aws_secrets.config.providers import BaseProvider
 from aws_secrets.config.providers.ssm.entry import SSMParameterEntry
 from aws_secrets.representers.literal import Literal
 
 
 class SSMProvider(BaseProvider):
+    """
+        AWS SSM Parameter Provider
+
+        This class handles the AWS SSM Parameter features:
+        - deploy the changes
+        - add new secrets
+        - update existing secrets
+        - decrypt/encrypt secrets
+
+        Args:
+            config (`ConfigReader`): configuration file reader handler
+
+        Attributes:
+            logger (`Logger`): logger instance
+            global_tags (`Dict[str, str]`): map of global tags of the config file
+            session (`Session`): boto3 session object
+            kms_arn (`str`): Main Kms ARN
+            config_data (`Dict[str, Any]`): configuration parsed YAML
+            secrets_data (`Dict[str, Any]`): secrets parsed YAML
+            entries (`List[SSMParameterEntry]`): entries list
+    """
+
     def __init__(self, config) -> None:
         super(SSMProvider, self).__init__(config)
 
-    def load_entries(self):
+    def load_entries(self) -> List[SSMParameterEntry]:
+        """
+            The SSM parameter entries
+            location: config YAML file `parameters` list property
+
+            Returns:
+                `List[SSMParameterEntry]`: list of entries
+        """
         result = []
         self.logger.debug('Loading SSM Parameter entries')
         for param in self._get_data_entries():
@@ -25,7 +56,10 @@ class SSMProvider(BaseProvider):
 
         return result
 
-    def decrypt(self):
+    def decrypt(self) -> None:
+        """
+            Decrypt the SSM parameter with `SecureString` type in the config file
+        """
         self.logger.debug('Decrypting SSM Parameter entries')
         for item in self._get_data_entries():
             item_obj = next((p for p in self.entries if p.name == item['name']))
@@ -37,11 +71,29 @@ class SSMProvider(BaseProvider):
                 else:
                     item['value'] = decrypted_value
 
-    def find(self, name) -> BaseEntry:
+    def find(self, name: str) -> Optional[SSMParameterEntry]:
+        """
+            Find an AWS SSM Parameter by name
+
+            Args:
+                name (`str`): entry name
+
+            Returns:
+                `SSMParameterEntry`, optional: parameter object or None
+        """
         self.logger.debug(f'Finding SSM Parameter entries by name "{name}"')
         return next((e for e in self.entries if e.name == name), None)
 
-    def add(self, data):
+    def add(self, data: Dict[str, Any]) -> SSMParameterEntry:
+        """
+            Add new AWS SSM Parameter
+
+            Args:
+                data (`Dict[str, Any]`): new parameter data
+
+            Returns:
+                `SSMParameterEntry`: parameter object
+        """
         self.logger.debug(f'Add "{data["name"]}" SSM Parameter entry')
         entry = SSMParameterEntry(
             session=self.session,
@@ -53,7 +105,13 @@ class SSMProvider(BaseProvider):
 
         return entry
 
-    def update(self, data):
+    def update(self, data: Dict[str, Any]) -> None:
+        """
+            Update an existing parameter
+
+            Args:
+                data (`Dict[str, Any]`): updated parameter data
+        """
         self.logger.debug(f'Updating "{data["name"]}" SSM Parameter entry')
 
         for idx, entry in enumerate(self.entries):
@@ -71,7 +129,20 @@ class SSMProvider(BaseProvider):
                 self.logger.debug('Updating entry data in the data entries list')
                 data_entries[idx] = data
 
-    def deploy(self, filter_pattern, dry_run, confirm):
+    def deploy(
+        self,
+        filter_pattern: Optional[str],
+        dry_run: bool,
+        confirm: bool
+    ) -> None:
+        """
+            Deploy all AWS SSM Paremeter changes
+
+            Args:
+                filter_pattern (`Optional[str]`): resource filter pattern
+                dry_run: (`bool`): dry run flag, just calculate the changes, but not apply them.
+                confirm: (`bool`): CLI confirmation prompt for the changes.
+        """
         any_changes = False
         for parameter in self.filter(filter_pattern):
             if self._deploy_parameter(parameter, dry_run, confirm):
@@ -83,9 +154,20 @@ class SSMProvider(BaseProvider):
     def _deploy_parameter(
         self,
         parameter: SSMParameterEntry,
-        dry_run,
-        confirm
-    ):
+        dry_run: bool,
+        confirm: bool
+    ) -> bool:
+        """
+            Deploy the parameter changes
+
+            Args:
+                parameter (`SSMParameterEntry`): parameter entry object
+                dry_run: (`bool`): dry run flag, just calculate the changes, but not apply them.
+                confirm: (`bool`): CLI confirmation prompt for the changes.
+
+            Returns:
+                `bool`: if changes are found.
+        """
         self.merge_tags(parameter)
         changes = parameter.changes()
 
@@ -108,10 +190,21 @@ class SSMProvider(BaseProvider):
     def _deploy_parameter_changes(
         self,
         parameter: SSMParameterEntry,
-        changes,
-        dry_run,
-        confirm
-    ):
+        changes: Dict[str, Any],
+        dry_run: bool,
+        confirm: bool
+    ) -> None:
+        """
+            Update an existing parameter on the AWS environment
+
+            If the non replaceable attributes are found, recreate the resource instead of update
+
+            Args:
+                parameter (`SSMParameterEntry`): parameter entry object
+                changes (`Dict[str, Any]`): map of changes
+                dry_run: (`bool`): dry run flag, just calculate the changes, but not apply them.
+                confirm: (`bool`): CLI confirmation prompt for the changes.
+        """
         self.print_resource_name('Parameter', parameter.name)
         self.print_changes(changes)
 
@@ -129,13 +222,31 @@ class SSMProvider(BaseProvider):
             if (has_non_replaceable_changes is False and confirm and click.confirm(confirm_msg)) or confirm is False:
                 parameter.update(changes)
 
-    def get_sensible_entries(self):
+    def get_sensible_entries(self) -> List[Dict[str, Any]]:
+        """
+            Get sensible entries, all parameters typed `SecureString`
+
+            Returns:
+                `List[Dict[str, Any]]`: list of sensible entries data
+        """
         return list(filter(lambda p: p['type'] == 'SecureString', self._get_data_entries()))
 
-    def _get_secrets_entries(self):
+    def _get_secrets_entries(self) -> List[Dict[str, Any]]:
+        """
+            Get secrets config file entries
+
+            Returns:
+                `List[Dict[str, Any]]`: list of parameters data
+        """
         return self.secrets_data.get('parameters', [])
 
-    def _get_data_entries(self):
+    def _get_data_entries(self) -> List[Dict[str, Any]]:
+        """
+            Get the parsed config YAML parameters
+
+            Returns:
+                `List[Dict[str, Any]]`: list of parameters
+        """
         if 'parameters' not in self.config_data:
             self.config_data['parameters'] = []