+++ aws_secrets/miscellaneous/session.py
@@ -1,4 +1,5 @@
 import boto3
+from botocore.session import Session
 
 global aws_region
 global aws_profile
@@ -6,6 +7,12 @@ aws_region = ''
 aws_profile = ''
 
 
-def session():
+def session() -> Session:
+    """
+        Get boto3 session object
+
+        Returns:
+            `Session`: boto3 session object
+    """
     return boto3.session.Session(
         region_name=aws_region, profile_name=aws_profile)