+++ aws_secrets/tags/__init__.py
@@ -1,2 +1,0 @@
-from .output_stack import OutputStackTag
-from .cmd import CmdTag