+++ aws_secrets/tags/cmd.py
@@ -1,17 +1,48 @@
+import subprocess
+
 import click
 import yaml
-import subprocess
 from aws_secrets.miscellaneous import session
 from aws_secrets.miscellaneous.cloudformation import get_output_value
+from yaml.dumper import SafeDumper
+from yaml.nodes import ScalarNode
 
 
 class CmdTag(yaml.YAMLObject):
+    """
+        Custom YAML tag class
+        !cmd tag executes the command in the operation system and use the output console as value
+
+        Providers:
+            The providers work as variables in the `!cmd` command to replace some values
+            - `cf`
+            - `session`
+            - `aws`
+
+        Examples:
+            >>> !cmd "echo 'hello'"
+            value: `'hello'`
+
+            >>> !cmd "echo '${cf:<stack>.<output-name>}'"
+            value: `'<stack-output-value'`
+
+        Args:
+            value (`str`): value after `!cmd` tag (e.g `!cmd some` the value will be `some`)
+
+        Attributes:
+            value (`str`): value after `!cmd` tag (e.g `!cmd some` the value will be `some`)
+    """
+
     yaml_tag = u'!cmd'
 
-    def __init__(self, value):
+    def __init__(self, value: str):
         self.value = value
 
-    def resolve_variables(self):
+    def resolve_variables(self) -> None:
+        """
+            Resolve variables in the tag value
+            Variable pattern: `${<something>}`
+        """
         if self.value.find("${") == -1:
             return
 
@@ -39,7 +70,24 @@ class CmdTag(yaml.YAMLObject):
 
         self.resolve_variables()
 
-    def resolve_aws_provider(self, value, default_value):
+    def resolve_aws_provider(self, value: str, default_value: str):
+        """
+            Resolve `aws` provider variables
+
+            Args:
+                value (`str`): raw command value
+                default_value (`str`): default value for the provider
+
+            Examples:
+                >>> !cmd "${aws:profile}"
+                "--profile <cli-profile>"
+
+                >>> !cmd "${aws:profile, <default-value>}"
+                "--profile <cli-profile or default-value>"
+
+                >>> !cmd "${aws:profile}"
+                "" If the `aws-secrets ... --profile <profile>` is not specified
+        """
         self.check_aws_properties(value)
 
         output_value = ''
@@ -59,7 +107,24 @@ class CmdTag(yaml.YAMLObject):
 
         self.resolve_value(output_value)
 
-    def resolve_session_provider(self, value, default_value):
+    def resolve_session_provider(self, value: str, default_value: str):
+        """
+            Resolve `session` provider variables
+
+            Args:
+                value (`str`): raw command value
+                default_value (`str`): default value for the provider
+
+            Examples:
+                >>> !cmd "--profile ${session:profile}"
+                "--profile <cli-profile>"
+
+                >>> !cmd "--profile ${session:profile, <default-value>}"
+                "--profile <cli-profile or default-value>"
+
+                >>> !cmd "--profile ${session:profile}"
+                "--profile " If the `aws-secrets ... --profile <profile>` is not specified
+        """
         self.check_aws_properties(value)
 
         output_value = default_value
@@ -71,7 +136,18 @@ class CmdTag(yaml.YAMLObject):
 
         self.resolve_value(output_value)
 
-    def resolve_cf_provider(self, value, default_value):
+    def resolve_cf_provider(self, value: str, default_value: str):
+        """
+            Resolve `cf` provider variables
+
+            Args:
+                value (`str`): raw command value
+                default_value (`str`): default value for the provider
+
+            Examples:
+                >>> !cmd "${cf:<stack>.<output>}"
+                "<output-value>"
+        """
         stack_name = value.split(".")[0]
         output_name = value.split(".")[1]
         output_value = get_output_value(session.session(), stack_name, output_name)
@@ -81,16 +157,34 @@ class CmdTag(yaml.YAMLObject):
 
         self.resolve_value(output_value)
 
-    def resolve_value(self, output_value):
+    def resolve_value(self, output_value: str):
+        """
+            Resolve variable value
+
+            Args:
+                output_value (`str`): resolved value
+        """
         self.value = self.value.replace(self.value[self.value.find("${"):self.value.find("}")+1], output_value)
 
-    def check_aws_properties(self, value):
+    def check_aws_properties(self, value: str):
+        """
+            Check AWS properties for the `aws` and `session` providers
+
+            Args:
+                value (`str`): tag value
+        """
         allowed_values = ['profile', 'region']
         if value not in allowed_values:
             raise RuntimeError(f'Property `{value}` is not supported, ' +
                                f'provider `session` just supports {allowed_values} properties')
 
-    def __repr__(self):
+    def __repr__(self) -> str:
+        """
+            Resolve the variables and execute the command in the operational system.
+
+            Returns:
+                `str` stdout output
+        """
         self.resolve_variables()
 
         click.echo(f"Running command: {self.value}")
@@ -101,9 +195,9 @@ class CmdTag(yaml.YAMLObject):
         return output_value
 
     @classmethod
-    def from_yaml(cls, loader, node):
+    def from_yaml(cls, _, node):
         return CmdTag(node.value)
 
     @classmethod
-    def to_yaml(cls, dumper, data):
+    def to_yaml(cls, dumper: SafeDumper, data) -> ScalarNode:
         return dumper.represent_scalar(cls.yaml_tag, data.value)